+++ src/main/javascript/bundles/app-form.js
@@ -2,7 +2,7 @@ import "../components/datepicker";
 import "../components/timepicker";
 
 import "../js/application/app-form";
-import "../js/application/day-length-selector";
+import "../js/application/app-form-day-length-change-listener";
 import "../components/back-button";
 import "../components/application-replacement-select";
